requirements.txt
pandas~=2.1.3
torch~=2.1.1
matplotlib~=3.8.2
seaborn~=0.13.0
Pillow~=10.1.0
numpy~=1.26.2
scikit-learn~=1.3.2